First commit for the refactoring of the activator agent.
